Move `SyncState` to `:core:data-model`
