+++ core/data-model/src/main/java/com/ivy/core/data/Attachment.kt
@@ -1,6 +1,6 @@
 package com.ivy.core.data
 
-import com.ivy.core.data.common.Syncable
+import com.ivy.core.data.sync.Syncable
 import java.time.LocalDateTime
 import java.util.*
