+++ core/data-model/src/main/java/com/ivy/core/data/Category.kt
@@ -1,6 +1,10 @@
 package com.ivy.core.data
 
-import com.ivy.core.data.common.*
+import com.ivy.core.data.common.Archiveable
+import com.ivy.core.data.common.ItemIconId
+import com.ivy.core.data.common.IvyColor
+import com.ivy.core.data.common.Reorderable
+import com.ivy.core.data.sync.Syncable
 import java.time.LocalDateTime
 import java.util.*
