+++ core/data-model/src/main/java/com/ivy/core/data/Tag.kt
@@ -3,7 +3,7 @@ package com.ivy.core.data
 import com.ivy.core.data.common.Archiveable
 import com.ivy.core.data.common.IvyColor
 import com.ivy.core.data.common.Reorderable
-import com.ivy.core.data.common.Syncable
+import com.ivy.core.data.sync.Syncable
 import java.time.LocalDateTime
 import java.util.*
