+++ core/data-model/src/main/java/com/ivy/core/data/sync/Syncable.kt
@@ -1,4 +1,4 @@
-package com.ivy.core.data.common
+package com.ivy.core.data.sync
 
 import java.time.LocalDateTime
