+++ core/persistence/src/main/java/com/ivy/core/persistence/api/data/Saveable.kt
@@ -1,5 +1,7 @@
 package com.ivy.core.persistence.api.data
 
+import com.ivy.core.data.sync.SyncState
+
 data class Saveable<T>(
     val item: T,
     val sync: SyncState,
